requirements.txt
pytest
setuptools
pydantic_settings
httpx~=0.28.1

pathlib~=1.0.1
ujson~=5.10.0
numpy~=2.2.6

uvicorn~=0.38.0
fastapi~=0.124.0
pydantic~=2.11.4
websockets~=15.0.1
jinja2~=3.1.6
python-multipart~=0.0.20
ujson
soundfile~=0.13.1

pydub~=0.25.1

tqdm~=4.67.1
psutil~=7.0.0


librosa~=0.11.0

# For punctuation
transformers~=4.51.3

# For diarisation
scikit-learn~=1.5.2
python_speech_features
umap-learn~=0.5.7
hdbscan~=0.8.40

# for VAD
requests~=2.32.3

# For monitoring
pynvml~=12.0.0
nvidia-ml-py~=12.570.86

aiofiles~=24.1.0

watchdog~=6.0.0

# For ASR
onnx-asr == 0.10.2
hf_xet
onnxruntime

starlette

# For Auth / JWT
pyjwt~=2.10.1
bcrypt~=4.3.0
